pyproject.toml
[tool.poetry]
name = "t5"
version = "0.1.0"
description = "T5 FastAPI Model"
authors = ["Nana Obeng"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.2"
fastapi = "^0.79.0"
python-dotenv = "^0.20.0"

torch = "^1.12.1"
sentencepiece = "^0.1.97"
numpy = "^1.23.1"
transformers = "^4.21.1"
isort = "^5.10.1"
mypy = "^0.971"
black = "^22.6.0"



[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^5.0.4"
flake8-bugbear = "^22.7.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
